Fix mobile menu
It's a bit tricky getting both anchors and plugin menu to work woth the same styles

Anchored elements are now continously repositioned even when not visible to prevent bug where it appears before it gets repositioned.
Input fields will retain focus when prefix of suffix buttons are pressed (this is a bit problematic when used together with ElementClickOnMouseDown where events come in a non normal order).